Added firebase authentication and firestore

+++ src/Redux/Categories/ActionCreator.js
@@ -1,7 +1,26 @@
+import { auth, firestore } from "../../Firebase";
 import * as ActionTypes from "./ActionTypes";
 
 export const createCategory = (category) => (dispatch) => {
-  dispatch(createOneCategory(category));
+  if (!auth.currentUser) {
+    return;
+  }
+
+  var userId = auth.currentUser.uid;
+
+  firestore
+    .collection("categories")
+    .doc(category._id)
+    .set({ ...category, userId })
+    .then((resp) => {
+      console.log(resp);
+      dispatch(createOneCategory(category));
+      dispatch(getCategories());
+      console.log("Category added");
+    })
+    .catch((err) =>
+      console.log("Err in Adding Category ACTION CREATOR", err.message)
+    );
 };
 
 const createOneCategory = (category) => ({
@@ -9,8 +28,23 @@ const createOneCategory = (category) => ({
   payload: category,
 });
 
-export const deleteCategory = (index) => (dispatch) => {
-  dispatch(deleteOneCategory(index));
+export const deleteCategory = (index, id) => (dispatch) => {
+  if (!auth.currentUser) {
+    return;
+  }
+
+  firestore
+    .collection("categories")
+    .doc(id)
+    .delete()
+    .then(() => {
+      console.log("Category deleted");
+      dispatch(getCategories());
+      dispatch(deleteOneCategory(index));
+    })
+    .catch((err) =>
+      console.log("Err in deleting category ACTION CREATOR", err.message)
+    );
 };
 
 const deleteOneCategory = (index) => ({
@@ -19,10 +53,60 @@ const deleteOneCategory = (index) => ({
 });
 
 export const editCategory = (category, index) => (dispatch) => {
-  dispatch(editOneCategory(category, index));
+  if (!auth.currentUser) {
+    return;
+  }
+
+  firestore
+    .collection("categories")
+    .doc(category.id)
+    .update(category)
+    .then(() => {
+      console.log("Category edited");
+      dispatch(editOneCategory(category, index));
+      dispatch(getCategories());
+    })
+    .catch((err) => {
+      console.log("Err in editing category", err.message);
+    });
 };
 
 const editOneCategory = (category, index) => ({
   type: ActionTypes.EDIT_CATEGORY,
   payload: { category, index },
+});
+
+export const getCategories = () => (dispatch) => {
+  if (!auth.currentUser) {
+    return;
+  }
+  dispatch(categoriesLoading());
+  firestore
+    .collection("categories")
+    .where("userId", "==", auth.currentUser.uid)
+    .get()
+    .then((resp) => {
+      let categoriesData = [];
+      resp.forEach((category) => {
+        const data = category.data();
+        categoriesData.push(data);
+      });
+      dispatch(categoriesSuccess(categoriesData));
+    })
+    .catch((err) => {
+      dispatch(categoriesErr(err.message));
+      console.log("Error in getting categories ACTION CREATOR", err.message);
+    });
+};
+
+const categoriesLoading = () => ({
+  type: ActionTypes.CATEGORIES_LOADING,
+});
+const categoriesErr = (message) => ({
+  type: ActionTypes.CATEGORIES_ERR,
+  payload: message,
+});
+const categoriesSuccess = (categories) => ({
+  type: ActionTypes.CATEGORIES_SUCCESS,
+  payload: categories,
 });
